# server/swagger.yaml
openapi: 3.0.0
info:
  title: sms
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /student/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - student
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - student
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
    post:
      summary: CRUD ACTION update
      tags:
        - student
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/student"
  /student/findByacademic_id/{key}:
    get:
      summary: CRUD ACTION findByacademic_id
      tags:
        - student
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
  /student/findByclasss_id/{key}:
    get:
      summary: CRUD ACTION findByclasss_id
      tags:
        - student
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
  /student/findBygender_id/{key}:
    get:
      summary: CRUD ACTION findBygender_id
      tags:
        - student
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
  /student/findByrole_id/{key}:
    get:
      summary: CRUD ACTION findByrole_id
      tags:
        - student
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
  /student/findByschool_id/{key}:
    get:
      summary: CRUD ACTION findByschool_id
      tags:
        - student
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
    
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    student:
      properties:
        _id:
          type: string
        address:
          type: string
        admission_date:
          type: integer
        admission_no:
          type: integer
        age:
          type: string
        birth_certificate:
          type: string
        caste:
          type: string
        date_of_birth:
          type: integer
        email:
          type: string
        father_address:
          type: string
        father_email:
          type: string
        father_name:
          type: string
        father_occupation:
          type: string
        father_phone:
          type: string
        first_name:
          type: string
        height:
          type: string
        last_name:
          type: string
        mobile:
          type: string
        mother_address:
          type: string
        mother_email:
          type: string
        mother_name:
          type: string
        mother_occupation:
          type: string
        mother_phone:
          type: string
        name:
          type: string
        national_id:
          type: string
        national_id_no:
          type: string
        roll_no:
          type: integer
        student_photo:
          type: string
        transfer_certificate:
          type: string
        weight:
          type: string
      required:
        -  _id